Move pipette visualizaer to its own file, show actual path along with path plans
